WIP: extracted spotify into a seperate service
