+++ web/modules/custom/music_search/music_search.services.yml
@@ -1,6 +1,9 @@
 services:
   music_search.service:
     class: Drupal\music_search\MusicSearchService
+    arguments: ["@config.factory", "@music_search_spotify.service"]
+  music_search_spotify.service:
+    class: Drupal\music_search\MusicSearchSpotifyService
     arguments: ["@config.factory"]
   music_search.artist:
     class: Drupal\music_search\Form\MusicSearchArtistForm
